refactor: move OwnerUserAuthenticationData interface to types.ts
