+++ src/app/network/types.ts
@@ -1,0 +1,11 @@
+export interface OwnerUserAuthenticationData {
+  token: string;
+  bridgeUser: string;
+  bridgePass: string;
+  encryptionKey: string;
+  bucketId: string;
+  // to manage B2B workspaces
+  workspaceId?: string;
+  workspacesToken?: string;
+  resourcesToken: string;
+}
